Updates tests for should_notify_* methods
