+++ keystone_api/apps/notifications/tests/test_models/test_Preference.py
@@ -1,17 +1,10 @@
 """Unit tests for the `Preference` class."""
 
-from datetime import date, datetime, timedelta
-from unittest.mock import Mock, patch
-
-from django.contrib.auth import get_user_model
 from django.test import TestCase
 
-from apps.allocations.factories import AllocationRequestFactory
 from apps.notifications.factories import PreferenceFactory
 from apps.notifications.models import default_expiry_thresholds, Preference
 from apps.users.factories import UserFactory
-
-User = get_user_model()
 
 
 class GetExpirationThresholdMethod(TestCase):
@@ -154,93 +147,3 @@ class SetUserPreferenceMethod(TestCase):
         Preference.set_user_preference(user=self.user, notify_on_expiration=False)
         preference.refresh_from_db()
         self.assertFalse(preference.notify_on_expiration)
-
-
-class ShouldNotifyPastExpirationMethod(TestCase):
-    """Test the determination of whether a notification should be issued for an expired allocation."""
-
-    @patch('apps.notifications.models.Notification.objects.filter')
-    def test_false_if_duplicate_notification(self, mock_notification_filter: Mock) -> None:
-        """Verify the return value is `False` if a notification has already been issued."""
-
-        mock_notification_filter.return_value.exists.return_value = True
-
-        request = AllocationRequestFactory(expire=date.today())
-        pref = PreferenceFactory(user=request.submitter, notify_on_expiration=True)
-        self.assertFalse(pref.should_notify_past_expiration(request.id))
-
-    def test_false_if_disabled_in_preferences(self) -> None:
-        """Verify the return value is `False` if expiry notifications are disabled in preferences."""
-
-        request = AllocationRequestFactory(expire=date.today())
-        pref = PreferenceFactory(user=request.submitter, notify_on_expiration=False)
-        self.assertFalse(pref.should_notify_past_expiration(request.id))
-
-    def test_true_if_new_notification(self) -> None:
-        """Verify the return value is `True` if a notification has not been issued yet."""
-
-        request = AllocationRequestFactory(expire=date.today())
-        pref = PreferenceFactory(user=request.submitter, notify_on_expiration=True)
-        self.assertTrue(pref.should_notify_past_expiration(request.id))
-
-
-class ShouldNotifyUpcomingExpirationMethod(TestCase):
-    """Test the determination of whether a notification should be issued for an upcoming expiration."""
-
-    def test_false_if_request_does_not_expire(self) -> None:
-        """Verify the return value is `False` if the request does not expire."""
-
-        request = AllocationRequestFactory(expire=None)
-        pref = PreferenceFactory(user=request.submitter, request_expiry_thresholds=[15])
-        self.assertFalse(
-            pref.should_notify_upcoming_expiration(request.id, request.expire)
-        )
-
-    def test_false_if_request_already_expired(self) -> None:
-        """Verify the return value is `False` if the request has already expired."""
-
-        request = AllocationRequestFactory(expire=date.today())
-        pref = PreferenceFactory(user=request.submitter, request_expiry_thresholds=[15])
-        self.assertFalse(
-            pref.should_notify_upcoming_expiration(request.id, request.expire)
-        )
-
-    def test_false_if_no_threshold_reached(self) -> None:
-        """Verify the return value is `False` if no threshold has been reached."""
-
-        request = AllocationRequestFactory(expire=date.today() + timedelta(days=15))
-        pref = PreferenceFactory(user=request.submitter, request_expiry_thresholds=[5])
-        self.assertFalse(
-            pref.should_notify_upcoming_expiration(request.id, request.expire)
-        )
-
-    def test_false_if_user_recently_joined(self) -> None:
-        """Verify the return value is `False` if the user is new."""
-
-        user = UserFactory(date_joined=datetime.now())
-        request = AllocationRequestFactory(expire=date.today() + timedelta(days=15))
-        pref = PreferenceFactory(user=user, request_expiry_thresholds=[15])
-        self.assertFalse(
-            pref.should_notify_upcoming_expiration(request.id, request.expire)
-        )
-
-    @patch('apps.notifications.models.Notification.objects.filter')
-    def test_false_if_duplicate_notification(self, mock_filter: Mock) -> None:
-        """Verify the return value is `False` if a notification has already been issued."""
-
-        mock_filter.return_value.exists.return_value = True
-
-        request = AllocationRequestFactory(expire=date.today() + timedelta(days=15))
-        pref = PreferenceFactory(user=request.submitter, request_expiry_thresholds=[5])
-        self.assertFalse(
-            pref.should_notify_upcoming_expiration(request.id, request.expire)
-        )
-
-    def test_true_if_new_notification(self) -> None:
-        """Verify the return value is `True` if a notification threshold has been hit."""
-
-        request = AllocationRequestFactory(expire=date.today() + timedelta(days=5))
-        pref = PreferenceFactory(user=request.submitter, request_expiry_thresholds=[15])
-        self.assertTrue(
-            pref.should_notify_upcoming_expiration(request.id, request.expire)
-        )
